added flag to delete collection

+++ holodeck/rag/pipeline.py
@@ -1,11 +1,11 @@
 from weaviate.collections import Collection
-import holodeck.rag.ollama_utils as ollama_utils
-import holodeck.rag.weaviate_utils as weaviate_utils
+import holodeck.ollama.ollama_utils as ollama_utils
+import holodeck.weaviate.weaviate_utils as weaviate_utils
 import holodeck.utilities.constants as constants
 from loguru import logger
 from unstructured.staging.base import convert_to_dict
 
-def pipeline_prep() -> None:
+def pipeline_prep(delete_collection: bool) -> None:
     
     logger.info("Getting embeddings client...")
     embeddingClient = ollama_utils.get_embeddings_client(constants.OLLAMA_LOCAL_URL)
@@ -16,8 +16,9 @@ def pipeline_prep() -> None:
     weaviateClient = weaviate_utils.create_weaviate_local_client()
     
     # adding this to delete basic chunking
-    logger.info("Deleting Weaviate collection...")
-    weaviate_utils.delete_collection(weaviateClient, constants.WEAVIATE_COLLECTION_NAME)
+    if delete_collection:
+        logger.info("Deleting Weaviate collection...")
+        weaviate_utils.delete_collection(weaviateClient, constants.WEAVIATE_COLLECTION_NAME)
     
     logger.info("Getting Weaviate collection...")
     weaviateCollection = weaviate_utils.get_collection(weaviateClient, constants.WEAVIATE_COLLECTION_NAME)
